requirements.txt
pandas==0.23.4
python-twitter==3.5